+++ apps/api/serializers/user_acquisitions.py
@@ -2,6 +2,9 @@ from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
+from pydantic import field_validator, Field
+from pydantic_core.core_schema import ValidationInfo
+
 from apps.api.serializers import Serializer
 from apps.api.serializers.entries import AcquisitionSerializer, EntrySerializer
 from apps.api.serializers.users import UserSerializer
@@ -15,8 +18,12 @@ class UserAcquisitionSerializer:
         range: Optional[str]
         user: UserSerializer.Minimal
         acquisition: AcquisitionSerializer.Nested
-        url: str
+        url: str = Field(default=None, validate_default=True)
         entry: EntrySerializer.Base
         expire_at: Optional[datetime]
         created_at: datetime
         updated_at: datetime
+
+        @field_validator("url", mode="before")
+        def generate_absolute_url(cls, v, info: ValidationInfo) -> Optional[UUID]:
+            return info.context["request"].build_absolute_uri(v)